backend/python/requirements.txt
fastapi==0.116.0
transformers==4.53.2
onnx==1.18.0
optimum==1.26.1
uvicorn==0.35.0
python-dotenv==1.1.1